backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.9
pandas==2.2.0
openpyxl==3.1.2
pydantic==2.6.1
pydantic-settings==2.1.0
langchain-core==0.1.18
langgraph==0.0.20
python-dotenv==1.0.1 